django_project_lab/memlife/memacc/serializers.py
from rest_framework import serializers
from .models import Profile, User


class ProfileSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Profile
        fields = ('id','date_of_birth', 'photo')
        
class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('first_name', 'last_name', 'email','password')
        