django_project_lab/requirements.txt
absl-py==0.12.0
anyio==3.1.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.1.5
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.3.0
blinker==1.4
blosc==1.7.0
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
coverage==5.5
cowsay==3.0
cryptography==3.4.7
cupshelpers==1.0
custom-serialiser==0.0.0
cycler==0.10.0
Cython==0.29.14
dask==2.8.1+dfsg
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-allauth==0.44.0
django-braces==1.14.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
fsspec==0.6.1
future==0.18.2
gast==0.3.3
geojson==2.5.0
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.4.1
importlib-metadata==1.5.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
jupyterlab-widgets==1.0.0
jwcrypto==0.8
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.2.1
nbclassic==0.3.1
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.2.0
numexpr==2.7.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.9
pandas==0.25.3
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
partd==1.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==2.9.0
plot==0.6.5
pluggy==0.13.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.6
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.20
pycups==1.9.73
Pygments==2.8.0
PyGObject==3.36.0
PyJWT==2.1.0
pylint==2.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyowm==3.1.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gflags==1.5.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2019.3
PyWavelets==0.5.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-image==0.16.2
scipy==1.6.1
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==1.9.5
sqlparse==0.4.1
sympy==1.7.1
systemd-python==234
tables==3.6.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
texttable==1.6.2
toml==0.10.2
toolz==0.9.0
tornado==6.1
traitlets==5.0.5
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.6
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
zipp==1.0.0
